Renamed controls so they're less confusing
