+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Xamls/DetailControl.xaml
@@ -40,6 +40,8 @@
         <RowDefinition
           Height="Auto" />
         <RowDefinition
+          Height="Auto" />
+        <RowDefinition
           Height="*" />
       </Grid.RowDefinitions>
 
@@ -177,8 +179,8 @@
         </Grid>
       </Border>
 
-      <nuget:PackageMetadataControl
-        x:Name="_packageMetadataControl"
+      <nuget:PackageDetailsTabControl
+        x:Name="_packageDetailsTabControl"
         Grid.Row="4"
         AutomationProperties.Name="{x:Static nuget:Resources.Accessibility_PackageMetadata}"
         Margin="0,12,0,0"
