+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Xamls/PackageReadmeControl.xaml
@@ -1,5 +1,5 @@
 <UserControl
-  x:Class="NuGet.PackageManagement.UI.PackageMetadataReadmeControl"
+  x:Class="NuGet.PackageManagement.UI.PackageReadmeControl"
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
